Move Dispatch calls into dispatch.py
